gh-163 creates and works with local k8s and no etcd

+++ example/deployment/local-docker/bash-scripts/dockerCleanSystem.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-. "$DIR/../../bash-scripts/setScriptPath.sh"
-"$DOCKERBASHSCRIPTS/dockerComposeDown.sh"
-docker system prune -a
+CLEANETCD=$1
+if [[ -n "$CLEANETCD" ]]; then
+    DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+    . "$DIR/../../bash-scripts/setScriptPath.sh"
+    "$DOCKERBASHSCRIPTS/dockerComposeDown.sh" "$CLEANETCD"
+    docker system prune -a
+else
+   echo "argument error"
+fi
